refactor(dashboard-rework): Moved the account query down
