+++ apps/website/src/app/page.tsx
@@ -1,12 +1,12 @@
-import { getChestCountPerAccount, getTotalAccounts } from "@/actions/account";
+import { getTotalAccounts } from "@/actions/account";
 import {
   getTotalChests,
   getLatestChest,
   getPeakOpeningHourData,
 } from "@/actions/chest";
 import { getChestCountPerEvent, getPossibleChestCount } from "@/actions/event";
+import AccountCard from "@/components/account-card";
 import CategoryCard from "@/components/category-card";
-import ChestCountAccountChart from "@/components/chest-count-account-chart";
 import ChestCountEventTable from "@/components/chest-count-event-table";
 import LatestChestCard from "@/components/latest-chest-card";
 import PeakOpeningHourCard from "@/components/peak-opening-hour-card";
@@ -19,7 +19,6 @@ export default async function Dashboard() {
   const chestCount = await getTotalChests();
   const accountCount = await getTotalAccounts();
   const possibleChestCount = await getPossibleChestCount();
-  const accounts = await getChestCountPerAccount();
   const chest = await getLatestChest();
   const events = await getChestCountPerEvent();
   const peakOpeningHourData = await getPeakOpeningHourData();
@@ -44,7 +43,7 @@ export default async function Dashboard() {
         <CategoryCard filters={filters} />
       </div>
       <RewardCard filters={filters} />
-      <ChestCountAccountChart accounts={accounts} />
+      <AccountCard filters={filters} />
       <ChestCountEventTable events={events} accountCount={accountCount} />
     </div>
   );
